base-files: use /dev/root in /etc/fstab for systemd support
follow openembedded-core commit c509f948d9c575c45af8c5ed1cb1692c0ca5dade

systemd does not recognize "rootfs" in /etc/fstab so the root
filesystem is not checked. As a result, the following message
is logged by journalctl:
systemd-fstab-generator[68]: Checking was requested for "rootfs", but it
is not a device

Changing "rootfs" to "/dev/root" in /etc/fstab allows systemd to
check the root filesystem when the kernel is booted with the root
filesystem mounted read-only.