connman: blacklist USB gadget devices
Connman gained support for USB gadget. The typical usecase is to use
USB gadget for tethering. In this use case the module would be the
device sharing the connection to the USB host connected to the module.

However, our typical use case in mind for RNDIS is a simple ethernet
connection between a development host and the module, hence no need
for tethering/routing. This however is not easily doable with connman,
since connman tries to have one connection online only. When enabling
USB gadget as a normal ethernet device, this typically leads to classic
ethernet not being configured since the USB RNDIS connection is already
up and running.

By blacklisting USB gadget, connman won't touch the device and we can
use other means of bringing up the network interface.