Revert "angstrom-lxde-image: use udisks only for automount"
This reverts commit a1040ca051f6c66bc5205502a462d55490653360.

http://cgit.openembedded.org/openembedded-core/commit/meta/recipes-core/systemd/systemd_219.bb?h=fido&id=0cd5a0d5445ff95a8ef4abe24fef705957935e81
fixes the automount issues we have during boot.
So revert the patch which completely disabled the feature.

This (re)introduces the issue that removable media attached during runtime
does get mounted with dbus / udisk in a different fashion.
But the fact that media connected at boot is actually mounted before the
system reaches its running state and that automount works when one disables
the start of the display manager outweights this.