update.sh: use moduletype specific folder for output files
With update.sh one creates files on e.g. a SD-card or a tftp server
which enables updating the SW on a module from U-Boot running on
said module.

This commit adds a module type specific directory on that output media
so that update data for multiple module types can be stored on one media.

Generic U-Boot scripts in the media's root directory do select the
matching update data depending on module type.