python: add a PACKAGECONFIG for _tkinter
Currently _tkinter is only built when tcl/tk is available before the
python do_configure task is built.

Make this deterministic by providing a PACKAGECONFIG for it.
If tkinter is in PACKAGECONFIG add the needed dependencies,
if not force that tkinter is not built.

Note that currently enabling tkinter does not work due to a cyclic
dependency:
'python' depends on 'tk/tcl' depends on 'cracklib' depends on 'python'

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>