trdx-extra: don't deploy license text
When building for a second machine after building for one the rootfs
creation fails because some license files are not found.

I belive this is caused by the following commit:
https://github.com/Angstrom-distribution/meta-angstrom/commit/dae094d69a1f8651c822fb35d606bf59a615f244

All packages which are created from sstate fail do install in the rootfs
likely because the license file is only available in the subdirectory of
the machine for which it was originally built.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>