update.sh: use test expressions in if statement for colibri-t20
Since we introduced set -e commands which fail will exit silently.
This leads to no error message when the grep which tests for
specified parameters failed.

We could solve it with || true, but it seems more lightweight to
just test for the strings inside the if statement. It also checks
for the full length (e.g. disallows "25" as RAM size).

The later fixed only 256 MB RAM size copy paste issue.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
Signed-off-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>