update.sh: imx7: use new U-Boot to flash modules
U-Boot delivered with V2.6 Alpha1 does not have SoC type detection
which can be used to load the device tree (imx7s/imx7d). Flash
U-Boot first and reset to make sure we use the new U-Boot to flash
the rest of the components.

The later corrected the commit message.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
Signed-off-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>