images: use packagegroup-fsl-gpu-libs
The old solution with using packagegroup-fsl-tools-gpu with a
bbappend to remove the big tools and examples needs more maintanance
and makes it hard for someone who really wants the examples or tools
in a image.

So change to a new packagegroup to deploy the libraries.
Also deploy OpenCL releated libraries seperately.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>
Acked-by: Stefan Agner <stefan.agner@toradex.com>