images: Remove rootfs size limitation of 4GB with update scripts
Our current update procedure requires the rootfs payload to fit in
one file on a FAT filesystem. This limits its usable size to 4GB.
Use the process of splitting up the files in 64MB chunks and then
doing the update which is also what the Ethernet update scripts
currently follow.

However due to the workaround for u-boot counting in hex and split
did in decimal, the number of chunks is limited to 89 resulting in
a maximum size of 5.5GB. We will remove the limitation in a follow
up patch.

Signed-off-by: Sanchayan Maity <sanchayan.maity@toradex.com>
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>