images: Remove update script limitation of 5GB rootfs
Our current update procedure reads the rootfs payload in pre
split files of 64 MB size. However due to the workaround for
u-boot counting in hex and split did in decimal, the number of
chunks is limited to 89 resulting in a maximum size of 5.5GB.

Update the u-boot counting procedure to count in 3 digits from
the current 2 digits. This allows more than 800 plus chunks
which should take care of rootfs size and any future eMMC
update requirements.

Signed-off-by: Sanchayan Maity <sanchayan.maity@toradex.com>
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>