tdx-image-fstype: tar with --numeric-owner
If --numeric-owner is not used when creating and/or extracting the archive
containing the rootfs then tar tries to change the numeric uid/gid of the
files based on owner/group names of the host used to extract the archive.

Create the archive with --numeric-owner to prevent this from happening when
extracting.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>
Acked-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>