images: make console image inherit from core-image
Some image features (e.g. tools-debug) do not currently work with the
console-tdx-image. The reason is that it is inheriting from the image
class instead of the core-image class, which defines such extra
features. Fortunately, this change doesn't need extra modifications in
the file. The documentation shows core-image inheritance as the standard
way to create new images, so we are advised to do it this way from now
on.

Signed-off-by: Gerard Salvatella <gerard.salvatella@toradex.com>
Acked-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>