image-fstype: rework files installation
This adds a new recipe tdx-updater which provides the files needed by
tdx-image-fstype.inc rather than copying them in a non OE way in
tdx-image-fstype.inc.

This required to restructure the layout in files. An additional update
directory is inserted. Also in each of the overrides directories a
symlink to library is created.

This now allows to build legacy update images with a custom machine
configuration if one adds the relevant machine override, e.g. for a
machine targeting a colibri-imx6 module:
MACHINEOVERRIDES .= ":colibri-imx6"

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>