images: add a new console/graphical/verification image
This is meant to replace the current demo images.

The console and graphical image can be used as a base to create tailored
images to the requirment of a specific use case.

The verification image is meant to test the BSP. It provides the tools
to exercise the CPU / GPU / VPU and the various SOC peripherals.

Related-to: ELB-1284
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>