images: base-tdx-cli: add set-hostname service
The service "set-hostname" adds a serial number of a module to a
machine hostname. It allows a customer to identify a module he
is working with simply.

Add it to the "packagegroup-base-tdx-cli", so that all Reference
Multimedia and Minimal images have hostnames like Torizon in the
style of "colibri-imx8x-06748675". Also, customers using the
"packagegroup-base-tdx-cli" inherit that and it lands in their
production image.

Related-to: ELB-1322
Signed-off-by: Oleksandr Suvorov <oleksandr.suvorov@toradex.com>
(cherry picked from commit c00e37b6324b16361ddc08abfdf39acd54d60ce6)