packagegroup-tdx-cli: Add missing mwifiexap to verdin-imx8mp
mwifiexap is required to create the uap0 device to be used for Wi-Fi AP
mode, this used to be included because of the `mx8`
RRECOMMENDS_packagegroup-machine-tdx-cli,
however after the addition of a new `verdin-imx8mp` it is no longer
there.

Add mwifiexap explicitly to verdin-imx8mp packagegroup to solve the
issue.

Related-to: ELB-4535
Fixes: 7df50f3ff17a ("packagegroup-tdx-cli: verdin-imx8mp: add btuart")
Signed-off-by: Francesco Dolcini <francesco.dolcini@toradex.com>
(cherry picked from commit ce3c1925df34b4d299b2dd1003ced41b9485ce41)