base-files: drop Toradex fstab
We dont have to maintain a Toradex specific fstab, it leads we miss the
latest updates in base base-files recipe.

The only change we really want is a have a extra line in fstab to mount
boot partition when system starts up, we can change to add one more udev
rule to create a symbolic link /boot ->/media/BOOT-PARTITION-MOUNTPOINT
in BSP layer instead, hence we can drop this fstab here.

Signed-off-by: Ming Liu <ming.liu@toradex.com>