wayland-app-launch: add xwayland_display env variable
Weston is no longer being started as root. To allow applications and
other users to start applications that render to weston there is the
socket-file in /run/wayland-0 created.

Do also specify the user we want the app to be started from. Do also
instruct systemd with PAMName to treat as a login so XDG_RUNTIME_DIR is
also set.

Related-to: ELB-4784
Signed-off-by: Philippe Schenker <philippe.schenker@toradex.com>