wayland-app-launch: remove shell-script
weston uses now systemd-notify to notify systemd when it is up and
running. This is the reason why we do no longer need
wayland-app-launch.sh with the synchronization point in it.

Remove wayland-app-launch.sh script. Remove also the init directory
feature as it is no longer needed. qtsmarthome works now from any
directory.

Related-to: ELB-4784
Signed-off-by: Philippe Schenker <philippe.schenker@toradex.com>