wayland-app-launch: do not restart on failure
If the wayland application fails to start we no longer restart it. This
can happen when there is no display connected. For example the
Qt5_CinematicExperience will fail to start with the following message:
xdg_surface geometry (1920 x 1080) is larger than the configured
fullscreen state (0 x 0)
If we keep restarting the application we spam the kernel log with
the following messages:
imx-drm display-subsystem: [drm] Cannot find any crtc or sizes

Related-to: ELB-4726
Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>