qt5: examples: cinematicexperience: Fix a crash with no screen
When there is no screen attached the Qt application fails to start.
The message when the application crashes is:
xdg_wm_base@17: error 4: xdg_surface geometry (500 x 500) is larger than the configured fullscreen state (0 x 0)
The Wayland connection experienced a fatal error: Protocol error

This leads to a kernel messages each time the application tries to
start. With BSP5 the application just tries to restart every time the
applications exits but doesn't spam the kernel log. The reason
why this changed is this new feature:
https://code.qt.io/cgit/qt/qtwayland.git/commit/?h=5.15&id=d58008c4310f99d0faebcfb2fd9aa9296b813ecf

The change of this commit fixes the application so that it properly
works with no screen attached and doesn't crash at all.

Related-to: ELB-4884
Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>